requirements.txt
tensorflow==1.3.0
pillow